+++ packages/hpl-nectar-staking/operations/stake.ts
@@ -1,3 +1,4 @@
+import { PROGRAM_ADDRESS } from "./../generated/index";
 import * as web3 from "@solana/web3.js";
 import * as splToken from "@solana/spl-token";
 import {
@@ -7,7 +8,7 @@ import {
   PROGRAM_ID,
 } from "../generated";
 import { AssetProof, AvailableNft } from "../types";
-import { PROGRAM_ID as BUBBLEGUM_PROGRAM_ID } from "@metaplex-foundation/mpl-bubblegum";
+import { MPL_BUBBLEGUM_PROGRAM_ID as BUBBLEGUM_PROGRAM_ID } from "@metaplex-foundation/mpl-bubblegum";
 import { PROGRAM_ID as AUTHORIZATION_PROGRAM_ID } from "@metaplex-foundation/mpl-token-auth-rules";
 import {
   getMetadataAccount_,