+++ packages/hpl-nectar-staking/pdas.ts
@@ -1,6 +1,6 @@
 import * as web3 from "@solana/web3.js";
 import { PROGRAM_ID } from "./generated";
-import { findProgramAddressSyncWithSeeds } from "@honeycomb-protocol/hive-control";
+import { PdaModule } from "@honeycomb-protocol/hive-control";
 
 /**
  * The `MetadataPDaType` type represents different types of Program Derived Accounts (PDAs)
@@ -67,7 +67,7 @@ export const getMetadataAccount_ = (
     }
   }
 
-  return findProgramAddressSyncWithSeeds(seeds, programId);
+  return PdaModule.findProgramAddressSyncWithSeeds(seeds, programId);
 };
 
 /**
@@ -87,7 +87,7 @@ export const getStakingPoolPda = (
   key: web3.PublicKey,
   programId: web3.PublicKey = PROGRAM_ID
 ) => {
-  return findProgramAddressSyncWithSeeds(
+  return PdaModule.findProgramAddressSyncWithSeeds(
     [Buffer.from("staking_pool"), project.toBuffer(), key.toBuffer()],
     programId
   );
@@ -110,7 +110,7 @@ export const getStakerPda = (
   wallet: web3.PublicKey,
   programId: web3.PublicKey = PROGRAM_ID
 ) => {
-  return findProgramAddressSyncWithSeeds(
+  return PdaModule.findProgramAddressSyncWithSeeds(
     [Buffer.from("staker"), wallet.toBuffer(), pool.toBuffer()],
     programId
   );
@@ -133,7 +133,7 @@ export const getNftPda = (
   mint: web3.PublicKey,
   programId: web3.PublicKey = PROGRAM_ID
 ) => {
-  return findProgramAddressSyncWithSeeds(
+  return PdaModule.findProgramAddressSyncWithSeeds(
     [Buffer.from("nft"), mint.toBuffer(), pool.toBuffer()],
     programId
   );
@@ -153,7 +153,7 @@ export const getDepositPda = (
   nftMint: web3.PublicKey,
   programId = PROGRAM_ID
 ) => {
-  return findProgramAddressSyncWithSeeds(
+  return PdaModule.findProgramAddressSyncWithSeeds(
     [Buffer.from("deposit"), nftMint.toBuffer()],
     programId
   );
@@ -173,7 +173,7 @@ export const getMultipliersPda = (
   pool: web3.PublicKey,
   programId = PROGRAM_ID
 ) =>
-  findProgramAddressSyncWithSeeds(
+  PdaModule.findProgramAddressSyncWithSeeds(
     [Buffer.from("multipliers"), pool.toBuffer()],
     programId
   );